Refactor telemetry structure
