+++ OpenTelemetry Toy Project/Features/Dice/Controllers/IDiceController.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace OpenTelemetry_Toy_Project.Features.Dice.Controllers;
+
+public interface IDiceController
+{
+    List<int> RollDice(string player, int? rolls);
+}
