--- OpenTelemetry Toy Project/IDiceController.cs
@@ -1,6 +1,0 @@
-namespace OpenTelemetry_Toy_Project;
-
-public interface IDiceController
-{
-    public List<int> RollDice(string player, int? rolls);
-}
